[ci skip] factorize `switch_plt_backend`
